+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [1.47.0] - 2022-11-25
 ### Added
 - Add "Dynamic cover carousel" component
+- Add func to apply alpha in DynamicCoverCarouselItem.
 
 ## [1.46.0] - 2022-11-09
 ### Changed